# src/main/java/com/agileengine/ecomm/generated/api/ProductApiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.agileengine.ecomm.generated.api;

import com.agileengine.ecomm.generated.model.Product;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "ProductApi", description = "the ProductApi API")
public interface ProductApiApi {

    /**
     * GET /products : Get a list of all products
     *
     * @return A list of products. (status code 200)
     */
    @Operation(
        operationId = "productsGet",
        summary = "Get a list of all products",
        tags = { "ProductApi" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of products.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Product.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<Product>> productsGet(
        
    ) throws Exception;


    /**
     * DELETE /products/{id} : Delete a product by ID
     *
     * @param id  (required)
     * @return Product deleted. (status code 204)
     */
    @Operation(
        operationId = "productsIdDelete",
        summary = "Delete a product by ID",
        tags = { "ProductApi" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Product deleted.")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/products/{id}"
    )
    
    ResponseEntity<Void> productsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) throws Exception;


    /**
     * GET /products/{id} : Get a product by ID
     *
     * @param id  (required)
     * @return A single product. (status code 200)
     */
    @Operation(
        operationId = "productsIdGet",
        summary = "Get a product by ID",
        tags = { "ProductApi" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A single product.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Product> productsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) throws Exception;


    /**
     * PUT /products/{id} : Update an existing product
     *
     * @param id  (required)
     * @param product  (required)
     * @return Product updated. (status code 200)
     */
    @Operation(
        operationId = "productsIdPut",
        summary = "Update an existing product",
        tags = { "ProductApi" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product updated.")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/products/{id}",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> productsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) throws Exception;


    /**
     * POST /products : Create a new product
     *
     * @param product  (required)
     * @return Product created. (status code 201)
     */
    @Operation(
        operationId = "productsPost",
        summary = "Create a new product",
        tags = { "ProductApi" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Product created.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> productsPost(
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) throws Exception;

}